pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.101.0"}
python-multipart = "^0.0.6"
authlib = "^1.2.1"
httpx = "^0.24.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
loguru = "^0.7.0"
rich = "^13.5.2"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"


[tool.poetry.group.lint.dependencies]
bandit = "^1.7.5"
pylama = {extras = ["lint"], version = "^8.4.1"}
mypy = "^1.5.0"
black = "^23.7.0"
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.types.dependencies]
types-pyyaml = "^6.0.12.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
